File tests moved to separate file
